Backend/requirements.txt
fastapi
uvicorn
groq
langchain
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
python-dotenv
pydantic
sentence-transformers